+++ test/proxy.js
@@ -1,8 +1,8 @@
 const lightwallet = require('eth-signer')
 const Proxy = artifacts.require('Proxy')
 const TestRegistry = artifacts.require('TestRegistry')
-const waitForTxReceipt = require('./waitForTxReceipt')
-const assertThrown = require('./assertThrown')
+const waitForTxReceipt = require('./utils/waitForTxReceipt')
+const assertThrown = require('./utils/assertThrown')
 const Promise = require('bluebird')
 const ethJSABI = require("ethjs-abi")
 web3.eth = Promise.promisifyAll(web3.eth)
